FOllowed Markus' suggestions
