+++ VersionedObject/JsonLdHelper.cs
@@ -108,87 +108,7 @@ namespace VersionedObject
             return graph.GetHash();
         }
 
-        private static readonly Func<IEnumerable<IRIReference>, Func<IRIReference, Func<IRIReference, PersistentEdge>>,
-                Func<(IEnumerable<JProperty>, IEnumerable<PersistentEdge>), JProperty,
-                    (IEnumerable<JProperty>, IEnumerable<PersistentEdge>)>>
-            reifyJToken =
-                (persistentIris, MakeEdge) => (acc, prop) =>
-                    prop.Name switch
-                    {
-                        "@id" or "@type" => (acc.Item1.Append(prop), acc.Item2),
-                        _ => @prop.Value switch
-                        {
-                            JObject obj =>
-                                @obj.SelectToken("@id") switch
-                                {
-                                    JValue id => persistentIris.Any(i => i.ToString().Equals(id.ToString())) switch
-                                    {
-                                        true => (acc.Item1, acc.Item2.Append(MakeEdge(prop.Name)(id.ToString()))),
-                                        false => ReifyObjectChild(acc, prop, obj.Properties().ReifyEdges(MakeEdge, persistentIris))
-                                    },
-                                    _ => ReifyObjectChild(acc, prop, obj.Properties().ReifyEdges(MakeEdge, persistentIris))
-                                },
-                            JValue val =>
-                                ReifyPropertyChild(acc, MakeEdge(prop.Name), prop, val, persistentIris),
-                            JArray vals =>
-                                ReifyPropertyArray(acc, MakeEdge(prop.Name), MakeEdge, prop, vals, persistentIris),
-                            _ => throw new Exception("Expected JObject, JValue or JArray")
-                        }
-                    };
-
-        /// <summary>
-        /// Removes any references to the persistentIris in the props argument, and adds all these to the list of edges
-        /// </summary>
-        /// <param name="props"></param>
-        /// <param name="persistentIris"></param>
-        /// <returns></returns>
-        public static (IEnumerable<JProperty> props, IEnumerable<PersistentEdge> edges) ReifyEdges(this IEnumerable<JProperty> props,
-                Func<IRIReference, Func<IRIReference, PersistentEdge>> MakeEdge,
-                IEnumerable<IRIReference> persistentIris) =>
-                    props.Aggregate(
-                        (new List<JProperty>(), new List<PersistentEdge>()),
-                        reifyJToken(persistentIris, MakeEdge),
-                        acc => (acc.Item1, acc.Item2)
-                    );
-
-
-        private static (IEnumerable<JProperty> props, IEnumerable<PersistentEdge> edges) ReifyPropertyArray((IEnumerable<JProperty> props, IEnumerable<PersistentEdge> edges) acc, Func<IRIReference, PersistentEdge> MakeEdgeFromObject, Func<IRIReference, Func<IRIReference, PersistentEdge>> MakeEdge, JProperty prop, JArray vals, IEnumerable<IRIReference> persistentIris)
-        {
-            var edges =
-                from v in (
-                    from v in vals
-                    where v != null
-                    select v
-                )
-                select (Value: v, External:
-                        from p in persistentIris
-                        where p.ToString().Equals(v.ToString())
-                        select p
-                    );
-
-            var externalEdges =
-                from edge in edges
-                where edge.External.Any()
-                select MakeEdgeFromObject(edge.Value.ToString());
-
-            var internalEdges =
-                from edge in edges
-                where !edge.External.Any()
-                select reifyJToken(persistentIris, MakeEdge)(acc, prop);
-
-            return (acc.props.Append(new JProperty(prop.Name, new JArray(internalEdges))), acc.edges.Union(externalEdges));
-        }
-
-        private static (IEnumerable<JProperty> props, IEnumerable<PersistentEdge> edges) ReifyPropertyChild((IEnumerable<JProperty> props, IEnumerable<PersistentEdge> edges) acc, Func<IRIReference, PersistentEdge> MakeEdge, JProperty prop, JValue val, IEnumerable<IRIReference> persistentIris)
-        {
-            if (persistentIris.Any(i => i.ToString().Equals(val.ToString())))
-                return (acc.props, acc.edges.Append(MakeEdge(val.ToString())));
-            return (acc.props.Append(prop), acc.edges);
-        }
-
-        static (IEnumerable<JProperty> props, IEnumerable<PersistentEdge> edges) ReifyObjectChild((IEnumerable<JProperty> props, IEnumerable<PersistentEdge> edges) acc, JProperty prop, (IEnumerable<JProperty> props, IEnumerable<PersistentEdge> edges) children) =>
-            (acc.props.Append(new JProperty(prop.Name, new JObject(children.props))), acc.edges.Union(children.edges));
-
+       
         public static IRIReference GetIRIReference(this JToken jsonld) =>
             new(jsonld.SelectToken("@id")?.ToString() ?? throw new InvalidJsonLdException($"No @id field in object {jsonld}"));
